+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- `WindowCreator` and `WindowBuilder` now support initial_position.
+
+### Breaking Changes
+
+- `WindowCreator` and `WindowBuilder` now use `Points` as the unit for
+  `initial_size`. When creating the window, we now request the logical size
+  rather than pixels, correspondingly.
 ## v0.1.0-dev.1
 
 ### Breaking Changes